zuul 层不通 是由于Security导致的，目前没有找到好的解决方案

+++ andy-server/server-user-center/src/main/kotlin/com/andy/service/servierusercenter/service/impl/IUserServiceImpl.kt
@@ -123,7 +123,7 @@ class IUserServiceImpl: IUserService {
         }
         userEntity.account = registerInfo.account
         userEntity.email = email!!
-        userEntity.tel = tel!!.toInt()
+        userEntity.tel = tel!!
         userEntity.password = passwordEncoder.encode(registerInfo.password)
         val userDetailsEntity = UserDetailsEntity()
         userDetailsEntity.nickName = registerInfo.nickName
@@ -234,8 +234,6 @@ class IUserServiceImpl: IUserService {
         userEntity.password = passwordEncoder.encode(password)
 
         val userInfo = userDao.save(userEntity)
-
-
         return UserDetailBean(
                 userInfo.id!!,
                 0,
